Added first version of FORT terminal
This unit is the interface with the user. It manages to show the user the characters pool from which they can type their password. Later version will correct certain problems present in the current version, such as:

1. Needing another keystroke to compute a letter typed.
2. Position of the characters pool
3. Adding or withdrawing from the characters pool letters that are not used in passwords

Future versions of FORT will feature a password handler and actions control. The former will compare the password typed by the user to grant or not further access. The latter will perform certain actions on the target files.

Built in Python for the fun of it. I will also build another version of this project in C++.